Refactor: 모달창 레이아웃/모달창 상태관리 구현 로직에 맞게 코드 수정
