+++ src/components/classroomModal/comment/ReplySection.tsx
@@ -1,7 +1,8 @@
-import React, { FC, use, useState } from "react";
-import Layout from "../Layout";
+import React, { FC } from "react";
 import CommentForm from "./CommentForm";
 import Comment from "./Comment";
+import Layout from "../common/Layout";
+import useClassroomModal from "@/hooks/useClassroomModal";
 
 interface Comment {
   text: string;
@@ -12,18 +13,19 @@ interface ReplySectionProps {
   role: string;
   activeComment: Comment | null;
   handleReply: (reply: string) => void;
-  handleBtn: () => void;
 }
 
 const ReplySection: FC<ReplySectionProps> = ({
   role,
   activeComment,
   handleReply,
-  handleBtn,
 }) => {
+  const { replyCommentModalOpen } = useClassroomModal();
+
   return (
-    activeComment && (
-      <Layout handleBtn={handleBtn}>
+    activeComment &&
+    replyCommentModalOpen && (
+      <Layout>
         <h2 className="text-2xl font-bold">상세보기</h2>
         <Comment
           admin={role}
@@ -46,11 +48,9 @@ const ReplySection: FC<ReplySectionProps> = ({
           ))}
         </ul>
         <CommentForm
-          isReply={true}
           handleComment={reply => {
             handleReply(reply);
           }}
-          handleClose={handleBtn}
         />
       </Layout>
     )
